requirements.txt
fastapi[all]
pandas==1.5.1
transformers==4.23.1
tensorflow_probability==0.17.0
tensorflow-cpu==2.9.1
pyarrow==9.0.0
openpyxl==3.0.10
opencv-python-headless==4.6.0.66
jinja2==3.1.2
MarkupSafe==2.1.1
ghostscript==0.7
camelot_py==0.10.1
fastapi_auth2==1.0.3
